src/functions/scoring/requirements.txt
# Cloud Functions Framework
functions-framework==3.*

# LINE Bot SDK
line-bot-sdk==3.12.0

# Google Cloud Libraries
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-cloud-logging==3.10.0

# Utilities
python-dotenv==1.0.1
